Swap terminal color 4 and 12 to fix dir colors
The directory color in the neovim terminal while true colors are enabled
does not match the color in non-true color mode: the color should be
blue, but it's a light grey.

The effected terminal color seems to be 12, which is set to the grey.
Change this to the blue, and change color 4 to the grey so all colors
are represented.

This fixes #62.