Add terminal cursor highlight
Inside a terminal buffer, highlight the terminal cursor in a different
color to distinguish between the terminal cursor position and the Vim
cursor position.

This option does not apply while in terminal mode, since the two cursors
are in the same position. The real benefit is to tell where your
terminal cursor is while *not* in terminal mode (aka, normal mode).

Since this is new and potentially confusing behavior, put it behind a
setting, and explain it in the README.